Add First Draft of Lecture 7 Notes
This commit contains a first draft that quickly hits each topic discussed in
lecture 7. There are many missing parts that need to be added including
visual representations of merge, rebase, branches, and tags, code for all
the commands, as well as many clarifications and explanations for various
topics/ideas/vocabulary used or brought up in this draft. Some of these are
highlighted in the markdown file with TODOs.