Adding changes to Day14Notes.md to match Soko's revision
This commit updates Day14Notes.md to match Soko's review. This commit
adds an introduction with relevant material from the prior class as
well as a conclusion that wraps up the lecture. It also fixes various
bolding issues that I missed on my original commits. This is a graded
requirenment of project 2 of CS340.