Removing the Lecture14_3-6.md file
This commit removes the Lecture14_3-6.md file that originally held the
lecture notes for lecture 14. I replaced it with Day14Notes.md. This
commits just gets rid of an unnecessary file as it had been replaced.
This should be the last commit to have the finished product of the
lecture 14 notes for project 2 of CS340.